requirements.txt
# TEE (Tessera Embeddings Explorer) - Dependencies
# Install: pip install -r requirements.txt

# =============================================================================
# CORE DEPENDENCIES (required for web app)
# =============================================================================

# Web server
django>=4.2
django-cors-headers>=4.0.0
waitress>=3.0.0

# Tile server still uses Flask
flask>=3.0.0
flask-cors>=4.0.0

# Geospatial processing
geotessera>=0.7.0
numpy>=1.24.0
rasterio>=1.3.0
Pillow>=10.0.0

# Tile server
rio-tiler>=6.0.0

# Similarity search
faiss-cpu>=1.7.0

# Vectorized computations
scipy>=1.10.0

# Authentication
bcrypt>=4.0.0

# HTTP client
requests>=2.28.0

# =============================================================================
# OPTIONAL: UMAP visualization (Panel 4 in advanced viewer)
# =============================================================================
umap-learn>=0.5.0

# =============================================================================
# OPTIONAL: Data download sources (only needed for creating new viewports)
# =============================================================================
# Uncomment if using Google Earth Engine:
# earthengine-api

# Uncomment if using Microsoft Planetary Computer:
# pystac-client
# planetary-computer
# odc-stac
# rioxarray

# =============================================================================
# UTILITIES
# =============================================================================
tqdm>=4.65.0
pyyaml>=6.0
